rfrp_project/requirements.txt
opencv-python
face_recognition
dlib
streamlit
streamlit-webrtc
python-pptx
psutil
pillow
pandas
pytest
av
